chore: separate x and y axes
