+++ src/components/react/landing-pages/data-visualization/components/internal/bar/SimpleBarChart.tsx
@@ -1,125 +1,13 @@
-import { BarChart } from '@mantine/charts';
 import '@mantine/charts/styles.css';
 import { Box, Flex, Text, Tooltip, rgba, useMantineTheme } from '@mantine/core';
 import { useElementSize, useMouse } from '@mantine/hooks';
-import { extent, rollup, scaleBand, scaleLinear, sum, type ScaleBand, type ScaleLinear } from 'd3';
-import { useCallback, useEffect, useMemo, useRef, useState } from 'react';
-import { findDiscreteValuesUsingDivideAndConquer } from 'src/components/react/landing-pages/data-visualization/utils';
+import { extent, rollup, scaleBand, scaleLinear, sum } from 'd3';
+import { useCallback, useEffect, useMemo, useState } from 'react';
 import type { ChartBoundaries } from '../../../types';
-
-const BAR_GAP = 4;
-
-function XAxisTextWithTooltip({ boundaries, value }: { value: string; boundaries: ChartBoundaries }) {
-  const ref = useRef<HTMLParagraphElement>(null);
-  const isOverflowing = useMemo(() => (ref.current?.scrollWidth ?? 0) > (ref.current?.clientWidth ?? 0), []);
-  return (
-    <Tooltip label={value} position="top-start" hidden={!isOverflowing} withArrow>
-      <Text size="xs" c="dimmed" w={boundaries.bottom} className="rotate-45 origin-top-left cursor-default" truncate="end" ref={ref}>
-        {value}
-      </Text>
-    </Tooltip>
-  );
-}
-
-function XAxis({
-  xScale,
-  categories,
-  svgDimensions,
-  boundaries,
-}: {
-  xScale: ScaleBand<string>;
-  categories: string[];
-  svgDimensions: { height: number; width: number };
-  boundaries: ChartBoundaries;
-}) {
-  const ticks = useMemo(
-    () =>
-      !svgDimensions.width
-        ? []
-        : categories.map((category, i) => ({
-            value: category,
-            xOffset: xScale.bandwidth() * i + xScale.bandwidth() / 2 - BAR_GAP / 2,
-          })),
-    [categories, svgDimensions.width, xScale],
-  );
-
-  return (
-    <g transform={`translate(${boundaries.left}, ${svgDimensions.height - boundaries.bottom})`}>
-      <path
-        d={['M', BAR_GAP * -2, 0, 'L', svgDimensions.width - boundaries.right - boundaries.left + BAR_GAP, 0].join(' ')}
-        style={{ stroke: 'var(--mantine-color-dimmed)' }}
-      />
-      {ticks.map(({ value, xOffset }) => (
-        <g key={value}>
-          <line x1={xOffset} x2={xOffset} y1={0} y2={4} style={{ stroke: 'var(--mantine-color-dimmed)' }} />
-          <foreignObject x={xOffset} y={8} className="overflow-visible">
-            <XAxisTextWithTooltip key={value} value={value} boundaries={boundaries} />
-          </foreignObject>
-        </g>
-      ))}
-    </g>
-  );
-}
-
-function YAxis({
-  yScale,
-  svgDimensions,
-  boundaries,
-}: {
-  yScale: ScaleLinear<number, number>;
-  svgDimensions: { height: number; width: number };
-  amountDomain: [number, number];
-  boundaries: ChartBoundaries;
-}) {
-  const yAxisPoints = useMemo(() => {
-    const [min, max] = yScale.domain();
-    const yStep = Math.pow(10, Math.floor(Math.log10(Math.abs(max - min))));
-    const output = [];
-    for (let i = min; i <= max; i += yStep) {
-      output.push(i);
-    }
-    return output;
-  }, [yScale]);
-
-  const ticks = useMemo(
-    () =>
-      !svgDimensions.height
-        ? []
-        : yAxisPoints.map((point, i) => ({
-            value: point.toFixed(2),
-            yOffset: ((svgDimensions.height - boundaries.bottom - boundaries.top) / (yAxisPoints.length - 1)) * (yAxisPoints.length - 1 - i),
-          })),
-    [boundaries.bottom, boundaries.top, svgDimensions.height, yAxisPoints],
-  );
-
-  return (
-    <g transform={`translate(${boundaries.left}, ${boundaries.top})`}>
-      <path
-        d={['M', BAR_GAP * -1, 0, 'L', BAR_GAP * -1, svgDimensions.height - boundaries.bottom].join(' ')}
-        style={{ stroke: 'var(--mantine-color-dimmed)' }}
-      />
-      {ticks.map(({ value, yOffset }, index) =>
-        value ? (
-          <g key={value}>
-            <line x1={-8} x2={0} y1={yOffset} y2={yOffset} style={{ stroke: 'var(--mantine-color-dimmed)' }} />
-            <line
-              x1={0}
-              x2={svgDimensions.width - boundaries.left - boundaries.right + BAR_GAP}
-              y1={yOffset}
-              y2={yOffset}
-              style={{ stroke: 'var(--mantine-color-dimmed)', strokeDasharray: index !== 0 ? '6 4' : 'unset' }}
-            />
-            <foreignObject x={boundaries.left * -1 - 16} y={yOffset - 8} className="overflow-visible">
-              <Text size="xs" c="dimmed" w={boundaries.left} ta="right" className="cursor-default">
-                {value}
-              </Text>
-            </foreignObject>
-          </g>
-        ) : null,
-      )}
-    </g>
-  );
-}
+import { XAxis } from './XAxis';
+import { YAxis } from './YAxis';
+import { BAR_GAP } from './constants';
+import { findDiscreteValuesUsingDivideAndConquer } from './helpers';
 
 export function SimpleBarChart<DataType extends Record<string, string | number>>({
   data,
@@ -262,7 +150,7 @@ export function SimpleBarChart<DataType extends Record<string, string | number>>
           )}
         </g>
       </svg>
-      <BarChart h={450} data={mantineBars} dataKey="category" series={[{ name: 'amount', color: 'blue.6' }]} tickLine="xy" tooltipAnimationDuration={200} />
+      {/* <BarChart h={450} data={mantineBars} dataKey="category" series={[{ name: 'amount', color: 'blue.6' }]} tickLine="xy" tooltipAnimationDuration={200} /> */}
     </Flex>
   );
 }
