+++ src/components/react/landing-pages/data-visualization/components/internal/bar/constants.ts
@@ -1,0 +1,1 @@
+export const BAR_GAP = 8;
