+++ src/components/react/landing-pages/data-visualization/components/internal/bar/helpers.ts
