--- src/components/react/landing-pages/data-visualization/utils/bar/index.ts
@@ -1,1 +1,0 @@
-export * from './helpers';
